Class Chromosome

This class is defined to summarize the enrichment analysis about a chromosome.

Objects from the Class

chromosome is defined to build an object of class Chromosome in order to compute an enrichment analysis. A Chromosome object contains the original data, a list of SNPs, some results and resampling data.

When created, an Chromosome object is "empty". readEnrichment initializes a Chromosome object with value from PLINK computation and user's files. In this step, only the fields "Data", "LD", "SNP" are filled. reSample fills the fields: Table, EnrichmentRatio, Z, PValue and Resampling of a Chromosome.

Note that if reSample is executed on an Chromosome every new resampling is added to the original ones, pre-existing statistics are erased and computed again with the new resampling set.

Slots

Data

[data.frame]: a data.frame with 6 columns ("SNP", "PVALUE", "CHR", "MAF", "eSNP", "xSNP"). Where "eSNP" and "xSNP" are logical columns defining the lists of SNPs (extended or not).

LD

[data.frame]: a data.frame which contains LD informations between SNPs (computed with writeLD or PLINK).

eSNP, xSNP

[SNP]: contain a EnrichSNP object (whith slots: SNP, Table, EnrichmentRatio, Z, PValue and Resampling) for a list of SNPs (eSNP) and an extended one (xSNP).

% \item{xSNP}{[SNP]: contains a \code{\linkS4class{EnrichSNP}} object (whith slots: SNP, Table, EnrichmentRatio, Z, PValue and Resampling) for an extended list of SNPs.}

Extends

Class EnrichSNP, directly.

Methods

chromosome(Data, LD, eSNP, xSNP):

Generate and initialize a new Chromosome object.

object["slotName"]:

Get the value of the field slotName.

object["slotName"]<-value:

Set value to the field slotName.

show(object):

Return the formatted values of Chromosome object.

% \item{print(object):}{Return the formatted values of \code{\linkS4class{Chromosome}} object.} % \item{computeER(object, sigThresh, mc.cores):}{Compute an Enrichment Ratio for \code{\linkS4class{Chromosome}} object (internal use only).} % \item{doLDblock(object, mc.cores):}{Compute LD block and update \code{Data} field (internal use only).} % \item{reSample(object, nSample, sigThresh, MAFpool, extendMethod, mc.cores):}{Compute P-Value based upon a resampling of SNPs (\code{eSNP} and \code{xSNP}) and update an \code{\linkS4class{Chromosome}} object.} % \item{excludeSNP(object, excludeFile, nSample, MAFpool, mc.cores):}{Exclude SNPs from eSNPs list and xSNPs list and compute a new enrichment analysis.} % \item{reset(object, "slotName"):}{Reset the field \code{slotName}.}

Note

Chromosome object is not intended to be used alone on this version.\ It is a part of the Enrichment object.

See also

Examples

Data <- structure(
    list(
        SNP = c("rs4970420", "rs3766178",
                "rs10910030", "rs10910036",
                "rs2234167", "rs6661861"),
        PVALUE = c(0.9244, 0.167, 0.01177, 0.4267, 0.9728, 0.4063),
        CHR = c(1, 1, 1, 1, 1, 1),
        POS = c(1106473, 1478180, 2035684, 2183754, 2494330, 3043121),
        MAF = c(0.2149, 0.3117, 0.374, 0.3753, 0.1565, 0.06101),
        eSNP = c(0, 1, 1, 0, 0, 0),
        xSNP = c(0, 1, 1, 0, 0, 0)
    ),
    .Names = c("SNP", "PVALUE", "CHR", "POS", "MAF", "eSNP", "xSNP"),
    row.names = c("rs4970420", "rs3766178",
                  "rs10910030", "rs10910036",
                  "rs2234167", "rs6661861"),
class = "data.frame")

toyChr <- chromosome(Data = Data)
show(toyChr)

#>      ~~~ Class: Chromosome ~~~
#>   ~ Data : (6x7)
#>                      SNP      PVALUE    CHR     POS     MAF    eSNP   xSNP  
#>       rs4970420   rs4970420   0.92440    1    1106473  0.2149    0      0   
#>       rs3766178   rs3766178   0.16700    1    1478180  0.3117    1      1   
#>       rs10910030  rs10910030  0.01177    1    2035684  0.3740    1      1   
#>       rs10910036  rs10910036  0.42670    1    2183754  0.3753    0      0   
#>       rs2234167   rs2234167   0.97280    1    2494330  0.1565    0      0   
#>         .....       .....      .....   .....   .....   .....   .....  ..... 
#> 
#>   ~ LD : NA
#>   ~ eSNP :
#>    - List : (2) rs3766178 rs10910030
#>    - Table : (2x2) NA
#>    - EnrichmentRatio : NA
#>    - Z               : NA
#>    - PValue          : NA
#>    - Resampling      : (0x5) 0
#> 
#>   ~ xSNP :
#>    - List : (2) rs3766178 rs10910030
#>    - Table : (2x2) NA
#>    - EnrichmentRatio : NA
#>    - Z               : NA
#>    - PValue          : NA
#>    - Resampling      : (0x5) 0
#> 

toyChr

#>      ~~~ Class: Chromosome ~~~
#>   ~ Data : (6x7)
#>                      SNP      PVALUE    CHR     POS     MAF    eSNP   xSNP  
#>       rs4970420   rs4970420   0.92440    1    1106473  0.2149    0      0   
#>       rs3766178   rs3766178   0.16700    1    1478180  0.3117    1      1   
#>       rs10910030  rs10910030  0.01177    1    2035684  0.3740    1      1   
#>       rs10910036  rs10910036  0.42670    1    2183754  0.3753    0      0   
#>       rs2234167   rs2234167   0.97280    1    2494330  0.1565    0      0   
#>         .....       .....      .....   .....   .....   .....   .....  ..... 
#> 
#>   ~ LD : NA
#>   ~ eSNP :
#>    - List : (2) rs3766178 rs10910030
#>    - Table : (2x2) NA
#>    - EnrichmentRatio : NA
#>    - Z               : NA
#>    - PValue          : NA
#>    - Resampling      : (0x5) 0
#> 
#>   ~ xSNP :
#>    - List : (2) rs3766178 rs10910030
#>    - Table : (2x2) NA
#>    - EnrichmentRatio : NA
#>    - Z               : NA
#>    - PValue          : NA
#>    - Resampling      : (0x5) 0
#> 

toyChr <- chromosome()
toyChr["Data"] <- Data
toyChr

#>      ~~~ Class: Chromosome ~~~
#>   ~ Data : (6x7)
#>                      SNP      PVALUE    CHR     POS     MAF    eSNP   xSNP  
#>       rs4970420   rs4970420   0.92440    1    1106473  0.2149    0      0   
#>       rs3766178   rs3766178   0.16700    1    1478180  0.3117    1      1   
#>       rs10910030  rs10910030  0.01177    1    2035684  0.3740    1      1   
#>       rs10910036  rs10910036  0.42670    1    2183754  0.3753    0      0   
#>       rs2234167   rs2234167   0.97280    1    2494330  0.1565    0      0   
#>         .....       .....      .....   .....   .....   .....   .....  ..... 
#> 
#>   ~ LD : NA
#>   ~ eSNP :
#>    - List : (0) NA
#>    - Table : (2x2) NA
#>    - EnrichmentRatio : NA
#>    - Z               : NA
#>    - PValue          : NA
#>    - Resampling      : (0x5) 0
#> 
#>   ~ xSNP :
#>    - List : (0) NA
#>    - Table : (2x2) NA
#>    - EnrichmentRatio : NA
#>    - Z               : NA
#>    - PValue          : NA
#>    - Resampling      : (0x5) 0
#> 

is.chromosome(toyChr) # TRUE

#> [1] TRUE