Class "EnrichSNP"

This class is defined to summarize the enrichment analysis. It's a part of Chromosome and Enrichment classes.

Slots

List

[vector(character)]: a list of SNPs used to compute enrichment (e.g. eSNP or xSNP).

Table

[matrix]: Contingency table with SNPs (columns) and P-Values from signal (rows).

EnrichmentRatio

[numeric]: Enrichment Ratio is computed on the contingency table (Table slot).

Z

[numeric]: A statistic computed from EnrichmentRatio and resampling results.

PValue

[numeric]: P-Value associated with the statistic Z.

Resampling

[matrix]: A matrix with by row, the contingency table and the odds ratio for each resampling.

Methods

% \item{enrichSNP(List, Table, EnrichmentRatio, Z, PValue, Resampling): }{Generate and initialize a new \code{\linkS4class{EnrichSNP}} object.}
object["slotName"]:

Get the value of the field slotName.

object["slotName"]<-value:

Set value to the field slotName.

show(object):

Return the formatted values of EnrichSNP object.

% \item{print(x): }{Return the values of \code{\linkS4class{EnrichSNP}} object as a matrix.}

Note

EnrichSNP object is not intended to be use directly by user. It is a part of the Enrichment and Chromosome object.

See also