Get all eSNP/xSNP which are enriched

getEnrichSNP get all eSNP/xSNP in a Enrichment object which are significant in the signal according to sigThresh defined in readEnrichment.

# S4 method for Enrichment
getEnrichSNP(object, type = "eSNP")

Arguments

object
[Enrichment]: an object of class Enrichment.

type
[character]: extract eSNP or xSNP data.

Value

Return a data.frame with eSNP/xSNP which are enriched in signal given to signalFile in function readEnrichment.

See also

Examples

# NOT RUN {
data(toyEnrichment)
eSNPenriched <- getEnrichSNP(object = toyEnrichment, type = "eSNP")
head(eSNPenriched)
# }