~ Internal: snpEnrichment objects and methods ~

These are not intended to be called by the user. If you have a specific need or for more details on snpEnrichment, feel free to contact the author.

.mclapply(X, FUN, mc.cores, ...)

.EnrichSNP.show(object)

.Chromosome.show(object)

.Enrichment.show(object)

.checkFilePath(path)

.checkSignalFile(signalFile)

.checkSnpInfoDir(snpInfoDir)

.checkSnpListDir(snpListDir, pattern)

.checkTranscript(data, transcriptFile, distThresh)

.compareEnrich(object1, object2, nSample, empiricPvalue, sigThresh,
  MAFpool, mc.cores)

.enrichmentRatio(table)

.readFiles(pattern, snpInfoDir, snpListDir, distThresh)

.readFreq(pattern, snpInfoDir)

.readLD(pattern, snpInfoDir, ldDir)

.readSignal(pattern)

.readSNP(pattern, snpListDir)

.readTranscript(transcriptFile)

.reSample(object, nSample, empiricPvalue, sigThresh, MAFpool, mc.cores)

.splitByChrom(pattern, snpListFile, directory)

.verbose(expr)

.writeFreq(pattern, snpInfoDir)

.writeSignal(pattern, snpInfoDir, signalFile)

GC(verbose = getOption("verbose"), reset = FALSE)

is.EnrichSNP(object)

is.chromosome(object)

is.enrichment(object)