Initialize files for enrichment analysis

initFiles create several files needed to run readEnrichment. ".frq" and ".signal" are created with PLINK. LD computation can be run with writeLD or with PLINK.

initFiles(pattern = "Chrom", snpInfoDir, signalFile, mc.cores = 1)

Arguments

pattern
A character string containing a expression to be matched with all chromosomes files (e.g.,"Chrom" for files which start by "Chrom" followed by the chromosome number).

snpInfoDir
character: character string naming a directory containing the reference data in a PLINK format (.bed, .bim and .fam).

signalFile
The name of the signal file which the data are to be read from (2 columns: "SNP" and "PVALUE"). Each row of the table appears as one line of the file. If it does not contain an _absolute_ path, the file name is _relative_ to the current working directory, getwd. The fields separator character have to be a space " " or a tabulation "\t".

mc.cores
numeric: the number of cores to use (default is 1), i.e., at most how many child processes will be run simultaneously. Must be at least one, and parallelization requires at least two cores.

Value

This function writes several files, in the temporary directory (defined in R_SESSION_TMPDIR), nothing else is returned. These files are used to build an Enrichment object by readEnrichment in order to compute enrichment analysis (reSample).

Examples

if (interactive()) {
  snpInfoDir <- system.file("extdata/snpInfo", package = "snpEnrichment")
  signalFile <- system.file("extdata/Signal/toySignal.txt", package = "snpEnrichment")
  initFiles(pattern = "Chrom", snpInfoDir, signalFile, mc.cores = 1)
}