Read and create EnrichmentRatio object

Read files created by initFiles and create an Enrichment object.

readEnrichment(pattern = "Chrom", signalFile,
               transcriptFile = "FALSE", snpListDir,
               snpInfoDir, distThresh = 1000,
               sigThresh = 0.05, LD = FALSE, ldDir = NULL,
               mc.cores = 1)

Arguments

pattern
[character]: character string containing a expression to be matched with all chromosomes files (e.g."Chrom" for files which start by "Chrom" followed by the chromosome number).

signalFile
[character]: the name of the signal file which the data are to be read from (2 columns: "SNP" and "PVALUE"). Each row of the table appears as one line of the file. If it does not contain an _absolute_ path, the file name is _relative_ to the current working directory, getwd. The fields separator character have to be a space " " or a tabulation "\t".

transcriptFile
[character or data.frame]: character string naming a file or a data.frame with four columns: Chromomosome, trancript's name, Starting and Ending positions. data(trancript) can be use as parameters. Default is FALSE.

snpListDir
[character]: character string naming a directory containing a list of SNPs for one or several chromosomes. snpListDir can be a single file with at least two columns: chromosome and rs name.

snpInfoDir
[character]: character string naming a directory containing the reference data in a PLINK format (*.bed, *.bim and *.fam).

distThresh
[numeric]: maximal distance (kb) between SNP and gene. distThresh is used if transcriptFile is set.

sigThresh
[numeric]: statistical threshold for signal (e.g. sigThresh = 0.05 for a given GWAS signal) used to compute an Enrichment Ratio.

LD
[logical]: LD=TRUE (default is FALSE) read LD compute with writeLD function or with PLINK. Note that, this setting can increase the computation's time, depending on number of SNPs in the signal file.

ldDir
[character]: character string naming a directory where the linkage disequilibrium files should be read (default ldDir=NULL is in temporary directory). LD files can be the LD output from plink.

mc.cores
[numeric]: the number of cores to use (default is mc.cores=1), i.e. at most how many child processes will be run simultaneously. Must be at least one, and parallelization requires at least two cores.

Value

Return an object of class Enrichment partly filled.

See also

Examples

# NOT RUN {
snpListDir <- system.file("extdata/List",
                          package = "snpEnrichment")
signalFile <- system.file("extdata/Signal/toySignal.txt",
                          package = "snpEnrichment")
snpInfoDir <- system.file("extdata/snpInfo", package = "snpEnrichment")
data(transcript)
transcriptFile <- transcript

initFiles(pattern = "Chrom", snpInfoDir, signalFile, mc.cores = 1)
toyData <- readEnrichment(pattern = "Chrom",
                         signalFile,
                         transcriptFile,
                         snpListDir,
                         snpInfoDir,
                         distThresh = 1000,
                         sigThresh = 0.05,
                         LD = FALSE,
                         ldDir = NULL,
                         mc.cores = 1)
toyData
# }